+++ script/MockPump.s.sol
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
-import "forge-std/Script.sol";
+import {Script, console} from "forge-std/Script.sol";
 import {MockPump} from "../mocks/pumps/MockPump.sol";
 
 // Script to deploy a {MockPump}.
 // Mockpump does not provide utility and is solely used for example.
 contract DeployMockPump is Script {
     function run() external {
-        uint256 deployerPrivateKey = vm.envUint("PRIVATE_KEY");
+        uint deployerPrivateKey = vm.envUint("PRIVATE_KEY");
         vm.startBroadcast(deployerPrivateKey);
 
         MockPump mockPump = new MockPump();
